HBSD: Add CA cert for binary update mechanism.
Introduce the CA cert for the binary update mechanism. This is done in
preparation before committing the script to build the update archive
and the script to apply the update archive.

Note that the CA private key can be stored locked away in a physical
vault. Since we're using an intermediate key for signing, the security
of the CA private key depends on the security of the physical holding
container.

Given the design of the update mechanism, if the intermediate private
key were compromised, the CA private key would be brought out of
storage to create a new intermediate key along with a CRL for the old.
The CA private key would then go immediately back in storage. This
same procedure would happen when the intermediate key expires.

The key was created with OpenSSL, using 4096bit RSA keys and SHA512 as
the hashing algorithm.

github-issue: #171
Signed-off-by: Shawn Webb <shawn.webb@hardenedbsd.org>

Sponsored-by: G2, Inc