HBSD: Only enable PIEified base for amd64 and i386.
Compiling (nearly) all of base causes issues with booting arm64 (and
likely arm, but I haven't verified that, yet). As I learn the arm64
architecture, and as I learn its boot process and what code is
involved with that, PIEified base will make its debut on arm64. I'm
hoping to complete that during BSDCan.

With this commit, PIEified base can be considered complete on amd64
and i386.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>